requirements.txt
anthropic>=0.39.0
openai>=1.50.0
APScheduler==3.11.0
alembic==1.16.1
authlib>=1.3.0
blinker==1.9.0
black==24.10.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.1.1
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-auth>=2.29.0
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.1.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pylint==3.3.3
pytest==8.3.4
pytest-cov==7.1.0
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.32.4
shapely==2.0.7
six==1.17.0
SQLAlchemy==2.0.31
structlog==23.1.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.2.2
Werkzeug==3.1.3
